adk/level5/requirements.txt

google-adk>=0.1.0
google-genai>=0.3.0
elevenlabs>=0.2.0
chromadb>=0.4.0
pydub>=0.25.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
aiofiles>=23.0.0
youtube-transcript-api>=0.6.0 